🔥 파일명 파일트리 타입스크립트 정리
