+++ src/components/Login/LoggedInStateBtn.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import styled from '@emotion/styled';
 import { useState } from 'react';
-import { COLOR } from '../../../constants';
+import { COLOR } from '../../constants';
+import { SpanProps } from 'src/types/Login';
 
 export const StayBtn = () => {
   const [loggedIn, setLoggedIn] = useState(false);
@@ -38,8 +39,3 @@ const Span = styled.div<SpanProps>`
   line-height: 15px;
   color: ${(props) => props.txtColor};
 `;
-
-//TS
-interface SpanProps {
-  txtColor: string;
-}
